Converted PrimeNodeJS to PrimeJavaScript
