--- PrimeBun/solution_1/run.sh
@@ -1,8 +1,0 @@
-#!/bin/sh
-bun run PrimeBun.js
-
-# yet to be implemented as (best I can tell) Bun
-# doesn't yet have cluster implemented
-# bun run PrimeBun_cluster.js
-
-bun run PrimeBun_memcopy.js
