--- PrimeBun/solution_2/package.json
@@ -1,6 +1,0 @@
-{
-  "devDependencies": {
-    "@types/node": "^18.0.3",
-    "typescript": "^4.7.4"
-  }
-}
