--- PrimeBun/solution_2/tsconfig.json
@@ -1,2 +1,0 @@
-{
-}
