--- PrimeNodeJS/solution_1/run.sh
@@ -1,4 +1,0 @@
-#!/bin/sh
-node PrimeNode.js
-node PrimeNode_cluster.js
-node PrimeNode_memcopy.js
